# Performance improvement for the views sys.sp_column_privileges_view and sys.sp_table_privileges_view