--- a/.github/configuration/upgrade-test-configuration.yml
+++ b/.github/configuration/upgrade-test-configuration.yml
@@ -101,7 +101,7 @@ upgrade-version: [{
 {
   upgrade-path: [
     {
-      version: '14.16',
+      version: '14.15',
       upgrade-type: null
     },
     { 
@@ -177,7 +177,7 @@ upgrade-version: [{
       upgrade-type: null
     },
     {
-      version: '14.16',
+      version: '14.15',
       upgrade-type: major
     },
     {