--- a/contrib/babelfishpg_tsql/sql/upgrades/babelfishpg_tsql--4.4.0--5.0.0.sql
+++ b/contrib/babelfishpg_tsql/sql/upgrades/babelfishpg_tsql--4.4.0--5.0.0.sql
@@ -681,6 +681,41 @@ LANGUAGE 'pltsql';
 
 GRANT EXECUTE on PROCEDURE sys.sp_helpuser TO PUBLIC;
 
+CREATE OR REPLACE FUNCTION sys.sp_column_privileges_func()
+RETURNS TABLE (
+    TABLE_QUALIFIER sys.sysname,
+    TABLE_OWNER sys.sysname,
+    TABLE_NAME sys.sysname,
+    COLUMN_NAME sys.sysname,
+    GRANTOR sys.sysname,
+    GRANTEE sys.sysname,
+    PRIVILEGE sys.varchar(32),
+    IS_GRANTABLE sys.varchar(3)
+) AS $$
+BEGIN
+    RETURN QUERY
+    SELECT
+        CAST(t2.dbname AS sys.sysname) AS TABLE_QUALIFIER,
+        CAST(s1.name AS sys.sysname) AS TABLE_OWNER,
+        CAST(t1.relname AS sys.sysname) AS TABLE_NAME,
+        CAST(COALESCE(SPLIT_PART(t6.attoptions[1], '=', 2), t5.column_name) AS sys.sysname) AS COLUMN_NAME,
+        CAST((SELECT orig_username FROM sys.babelfish_authid_user_ext WHERE rolname = t5.grantor::name) AS sys.sysname) AS GRANTOR,
+        CAST((SELECT orig_username FROM sys.babelfish_authid_user_ext WHERE rolname = t5.grantee::name) AS sys.sysname) AS GRANTEE,
+        CAST(t5.privilege_type AS sys.varchar(32)) COLLATE sys.database_default AS PRIVILEGE,
+        CAST(t5.is_grantable AS sys.varchar(3)) COLLATE sys.database_default AS IS_GRANTABLE
+    FROM pg_catalog.pg_class t1 
+        JOIN sys.pg_namespace_ext t2 ON t1.relnamespace = t2.oid
+        JOIN sys.schemas s1 ON s1.schema_id = t1.relnamespace
+        JOIN information_schema.column_privileges t5 ON t1.relname = t5.table_name AND t2.nspname = t5.table_schema
+        JOIN pg_attribute t6 ON t6.attrelid = t1.oid AND t6.attname = t5.column_name;
+END;
+$$ LANGUAGE plpgsql;
+
+CREATE OR REPLACE VIEW sys.sp_column_privileges_view AS
+SELECT *
+FROM sys.sp_column_privileges_func()
+WHERE GRANTEE NOT IN ('db_datareader', 'db_datawriter');
+
 CREATE OR REPLACE PROCEDURE sys.sp_column_privileges(
     "@table_name" sys.sysname,
     "@table_owner" sys.sysname = '',
@@ -700,7 +735,6 @@ BEGIN
 		IF EXISTS ( 
 			SELECT * FROM sys.sp_column_privileges_view 
 			WHERE pg_catalog.lower(@table_name) = pg_catalog.lower(table_name) and pg_catalog.lower(SCHEMA_NAME()) = pg_catalog.lower(table_qualifier)
-			AND GRANTEE NOT IN ('db_datawriter', 'db_datareader')
 			)
 		BEGIN 
 			SELECT 
@@ -735,7 +769,6 @@ BEGIN
 				AND (pg_catalog.lower('dbo')= pg_catalog.lower(table_owner))
 				AND ((SELECT COALESCE(@table_qualifier,'')) = '' OR pg_catalog.lower(table_qualifier) = pg_catalog.lower(@table_qualifier))
 				AND ((SELECT COALESCE(@column_name,'')) = '' OR pg_catalog.lower(column_name) LIKE pg_catalog.lower(@column_name))
-				AND GRANTEE NOT IN ('db_datawriter', 'db_datareader')
 			ORDER BY table_qualifier, table_owner, table_name, column_name, privilege, grantee;
 		END
 	END
@@ -755,13 +788,58 @@ BEGIN
 			AND ((SELECT COALESCE(@table_owner,'')) = '' OR pg_catalog.lower(table_owner) = pg_catalog.lower(@table_owner))
 			AND ((SELECT COALESCE(@table_qualifier,'')) = '' OR pg_catalog.lower(table_qualifier) = pg_catalog.lower(@table_qualifier))
 			AND ((SELECT COALESCE(@column_name,'')) = '' OR pg_catalog.lower(column_name) LIKE pg_catalog.lower(@column_name))
-			AND GRANTEE NOT IN ('db_datawriter', 'db_datareader')
 		ORDER BY table_qualifier, table_owner, table_name, column_name, privilege, grantee;
 	END
 END; 
 $$
 LANGUAGE 'pltsql';
 GRANT EXECUTE ON PROCEDURE sys.sp_column_privileges TO PUBLIC;
+
+CREATE OR REPLACE FUNCTION sys.sp_table_privileges_func()
+RETURNS TABLE (
+    TABLE_QUALIFIER sys.sysname,
+    TABLE_OWNER sys.sysname,
+    TABLE_NAME sys.sysname,
+    GRANTOR sys.sysname,
+    GRANTEE sys.sysname,
+    PRIVILEGE sys.sysname,
+    IS_GRANTABLE sys.sysname
+) AS $$
+BEGIN
+    RETURN QUERY
+    SELECT
+        CAST(t2.dbname AS sys.sysname) AS TABLE_QUALIFIER,
+        CAST(s1.name AS sys.sysname) AS TABLE_OWNER,
+        CAST(t1.relname AS sys.sysname) AS TABLE_NAME,
+        CAST((SELECT orig_username FROM sys.babelfish_authid_user_ext WHERE rolname = t4.grantor) AS sys.sysname) AS GRANTOR,
+        CAST((SELECT orig_username FROM sys.babelfish_authid_user_ext WHERE rolname = t4.grantee) AS sys.sysname) AS GRANTEE,
+        CAST(t4.privilege_type AS sys.sysname) AS PRIVILEGE,
+        CAST(t4.is_grantable AS sys.sysname) AS IS_GRANTABLE
+    FROM pg_catalog.pg_class t1 
+        JOIN sys.pg_namespace_ext t2 ON t1.relnamespace = t2.oid
+        JOIN sys.schemas s1 ON s1.schema_id = t1.relnamespace
+        JOIN information_schema.table_privileges t4 ON t1.relname = t4.table_name
+    WHERE t4.privilege_type = 'DELETE';
+END;
+$$ LANGUAGE plpgsql;
+
+CREATE OR REPLACE VIEW sys.sp_table_privileges_view AS
+-- Will use sp_column_priivleges_view to get information from SELECT, INSERT and REFERENCES (only need permission from 1 column in table)
+SELECT DISTINCT
+CAST(TABLE_QUALIFIER AS sys.sysname) COLLATE sys.database_default AS TABLE_QUALIFIER,
+CAST(TABLE_OWNER AS sys.sysname) AS TABLE_OWNER,
+CAST(TABLE_NAME AS sys.sysname) COLLATE sys.database_default AS TABLE_NAME,
+CAST(GRANTOR AS sys.sysname) AS GRANTOR,
+CAST(GRANTEE AS sys.sysname) AS GRANTEE,
+CAST(PRIVILEGE AS sys.sysname) COLLATE sys.database_default AS PRIVILEGE,
+CAST(IS_GRANTABLE AS sys.sysname) COLLATE sys.database_default AS IS_GRANTABLE
+FROM sys.sp_column_privileges_view
+
+UNION 
+-- We need these set of joins only for the DELETE privilege
+SELECT *
+FROM sys.sp_table_privileges_func()
+WHERE GRANTEE != 'db_datawriter';
 
 CREATE OR REPLACE PROCEDURE sys.sp_table_privileges(
 	"@table_name" sys.nvarchar(384),
@@ -789,7 +867,6 @@ BEGIN
 		IS_GRANTABLE FROM sys.sp_table_privileges_view
 		WHERE pg_catalog.lower(TABLE_NAME) LIKE pg_catalog.lower(@table_name)
 			AND ((SELECT COALESCE(@table_owner,'')) = '' OR pg_catalog.lower(TABLE_OWNER) LIKE pg_catalog.lower(@table_owner))
-			AND GRANTEE NOT IN ('db_datawriter', 'db_datareader')
 		ORDER BY table_qualifier, table_owner, table_name, privilege, grantee;
 	END
 	ELSE 
@@ -804,7 +881,6 @@ BEGIN
 		IS_GRANTABLE FROM sys.sp_table_privileges_view
 		WHERE pg_catalog.lower(TABLE_NAME) = pg_catalog.lower(@table_name)
 			AND ((SELECT COALESCE(@table_owner,'')) = '' OR pg_catalog.lower(TABLE_OWNER) = pg_catalog.lower(@table_owner))
-			AND GRANTEE NOT IN ('db_datawriter', 'db_datareader')
 		ORDER BY table_qualifier, table_owner, table_name, privilege, grantee;
 	END