--- a/contrib/babelfishpg_tsql/sql/upgrades/babelfishpg_tsql--4.4.0--5.0.0.sql
+++ b/contrib/babelfishpg_tsql/sql/upgrades/babelfishpg_tsql--4.4.0--5.0.0.sql
@@ -719,7 +719,7 @@ SELECT
     CAST(COLUMN_NAME AS sys.sysname),
     CAST(GRANTOR AS sys.sysname),
     CAST(GRANTEE AS sys.sysname),
-    CAST(PRIVILEGE AS sys.varchar(32)).
+    CAST(PRIVILEGE AS sys.varchar(32)),
     CAST(IS_GRANTABLE AS sys.varchar(3))
 FROM sys.sp_column_privileges_func()
 WHERE GRANTEE NOT IN ('db_datareader', 'db_datawriter');