--- a/test/JDBC/input/BABEL-SP_TABLE_PRIVILIGES-vu-verify.sql
+++ b/test/JDBC/input/BABEL-SP_TABLE_PRIVILIGES-vu-verify.sql
@@ -1,4 +1,3 @@
--- sla 160000
 use babel_sp_table_priviliges_vu_prepare_db1
 go