… test/python/expected/sql_validation_framework/expected_create.out
… test/python/expected/upgrade_validation/expected_dependency.out