### Description

1. Create internal functions for the views sys.sp_column_privileges_view and sys.sp_table_privileges_view to improve the performance by analysing the query plan. The performance degradation was caused by inaccurate row estimates resulting from an extra JOIN with babelfishpg_authid_user_ext table, leading to suboptimal query plans.

3. Fixed upgrade workflow by using correct source versions.


### Issues Resolved

Task: BABEL-5353
Signed-off-by: Shalini Lohia <lshalini@amazon.com>

### Test Scenarios Covered ###

Performance improvement:
1. JDBC Testing
 * before
```
BABEL-SP_COLUMN_PRIVILEGES: Failed! (62072/40000ms TIME OUT)
BABEL-SP_TABLE_PRIVILIGES-vu-verify: Failed! (42451/40000ms TIME OUT)
```
 * after:
```
BABEL-SP_COLUMN_PRIVILEGES:  Passed! (12431/40000ms OK)
BABEL-SP_TABLE_PRIVILIGES-vu-verify: Passed! (5025/40000ms OK)
```
2. Performance Testing

| query | Execution time before (ms) | Execution time after (ms)|
|-------| -------| ------|
|select * from sp_column_privileges_view | 4366 | 1408|
|select * from  sp_table_privileges_view |  5136.699 | 2231|
|EXEC sp_column_privileges @table_name = 't1'; | 5136.699 |1447.791|
|EXEC sp_table_privileges @table_name = 't1'; | 5195.126 | 2208.605|